app/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
python-dotenv>=1.0.0
openai>=1.3.0
pydantic>=2.0.0
websockets>=11.0.0
pyaudio>=0.2.13
ffmpeg-python>=0.2.0
python-multipart>=0.0.6